+++ cmd/frontend/internal/httpapi/router/router.go
@@ -43,6 +43,7 @@ const (
 	ReposUpdateMetadata    = "internal.repos.update-metadata"
 	Configuration          = "internal.configuration"
 	ExternalServiceConfigs = "internal.external-services.configs"
+	ExternalServicesList   = "internal.external-services.list"
 )
 
 // New creates a new API router with route URL pattern definitions but
@@ -98,6 +99,7 @@ func NewInternal(base *mux.Router) *mux.Router {
 	base.Path("/git/{RepoName:.*}/git-upload-pack").Methods("POST").Name(GitUploadPack)
 	base.Path("/phabricator/repo-create").Methods("POST").Name(PhabricatorRepoCreate)
 	base.Path("/external-services/configs").Methods("POST").Name(ExternalServiceConfigs)
+	base.Path("/external-services/list").Methods("POST").Name(ExternalServicesList)
 	base.Path("/repos/create-if-not-exists").Methods("POST").Name(ReposCreateIfNotExists)
 	base.Path("/repos/inventory-uncached").Methods("POST").Name(ReposInventoryUncached)
 	base.Path("/repos/inventory").Methods("POST").Name(ReposInventory)